refactor(plugins): separate conversation binding approval state
